refactor(croquis): move script_parser tests into submodule
